chore: handle content for credentials expiring alerts
